Quieten compiler warnings.

+++ interpreter/lex.c
@@ -102,7 +102,7 @@ LexInit(Picoc * pc)
 
 	TableInitTable(&pc->ReservedWordTable, &pc->ReservedWordHashTable[0], sizeof(ReservedWords) / sizeof(struct ReservedWord) * 2, TRUE);
 
-	for (Count = 0; Count < sizeof(ReservedWords) / sizeof(struct ReservedWord); Count++)
+	for (Count = 0; (unsigned long) Count < sizeof(ReservedWords) / sizeof(struct ReservedWord); Count++)
 	{
 		TableSet(pc, &pc->ReservedWordTable, TableStrRegister(pc, ReservedWords[Count].Word), (struct Value *)&ReservedWords[Count], NULL, 0, 0);
 	}
@@ -124,7 +124,7 @@ LexCleanup(Picoc * pc)
 
 	LexInteractiveClear(pc, NULL);
 
-	for (Count = 0; Count < sizeof(ReservedWords) / sizeof(struct ReservedWord); Count++)
+	for (Count = 0; (unsigned long) Count < sizeof(ReservedWords) / sizeof(struct ReservedWord); Count++)
 		TableDelete(pc, &pc->ReservedWordTable, TableStrRegister(pc, ReservedWords[Count].Word));
 }
